#71 Move I/O functions to io.go and update DumpSelectiveTables to write per line
